dashboardAdmin: products table fixed state bug
