+++ tsconfig.json
@@ -24,6 +24,6 @@
       "~/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/components/componetsDashboard/Cards/CardBarChart.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/components/componetsDashboard/Cards/Not used/CardBarChart.jsx"],
   "exclude": ["node_modules"]
 }
